Update DiningPhilosophersProblem.py
Implement the Dining Philosophers Problem. Philosophers are arranged in a circular list.
A philosopher that has finished eating is removed from the list. The rules:

1. Eating is done with a fix duration say 5secs.
2. Eating can only commence when both left and right forks are successfully acquired
3. The former fork is released if the latter fork cannot be acquired.
4. The problem concludes when all philosophers have dined.

Be creative in logging the actions of each Philosopher thread.  Logs must be saved to a file
via Python logging module.

In this exercise, you will learn:
1. How to use a thread
2. How to use a mutex to enforce thread synchronization
3. Use of python logging